flask-app/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.2
python-dotenv==1.0.1
bcrypt==4.1.3
PyJWT==2.8.0
requests==2.32.3
google-generativeai==0.7.2
pillow==10.4.0
pytesseract==0.3.10
certifi==2024.7.4
inference-sdk==0.9.17
